pyproject.toml
[project]
name = "caruaru-sgem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django>=6.0.3",
    "django-crispy-forms>=2.6",
    "django-filter>=25.2",
    "pillow>=12.1.1",
    "weasyprint>=68.1",
]
